+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentgetconfigop.py
@@ -1476,13 +1476,13 @@ class DeploymentGetConfigFunction(BaseModel):
     """
 
 
-class DeploymentGetConfigToolsTypedDict(TypedDict):
+class ToolsTypedDict(TypedDict):
     type: DeploymentGetConfigDeploymentsType
     r"""The type of the tool. Currently, only `function` is supported."""
     function: DeploymentGetConfigFunctionTypedDict
 
 
-class DeploymentGetConfigTools(BaseModel):
+class Tools(BaseModel):
     type: DeploymentGetConfigDeploymentsType
     r"""The type of the tool. Currently, only `function` is supported."""
 
@@ -1505,7 +1505,7 @@ class DeploymentGetConfigResponseBodyTypedDict(TypedDict):
     r"""Model Parameters: Not all parameters apply to every model"""
     type: NotRequired[DeploymentGetConfigType]
     r"""The type of the model. Current `chat`,`completion` and `image` are supported"""
-    tools: NotRequired[List[DeploymentGetConfigToolsTypedDict]]
+    tools: NotRequired[List[ToolsTypedDict]]
     r"""A list of tools the model may call. Currently, only functions are supported as a tool. Use this to provide a list of functions the model may generate JSON inputs for."""
 
 
@@ -1532,5 +1532,5 @@ class DeploymentGetConfigResponseBody(BaseModel):
     type: Optional[DeploymentGetConfigType] = None
     r"""The type of the model. Current `chat`,`completion` and `image` are supported"""
 
-    tools: Optional[List[DeploymentGetConfigTools]] = None
+    tools: Optional[List[Tools]] = None
     r"""A list of tools the model may call. Currently, only functions are supported as a tool. Use this to provide a list of functions the model may generate JSON inputs for."""