+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JSH5F6W9A2QZS7YBHE8CNFXD"
+        "01JSK8HMVXP84N54KS9XEG08HW"
     )
     r"""The id of the resource"""